requirements.txt
asgiref==3.8.1
Django==5.0.4
django-js-asset==2.2.0
django-mptt==0.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
